style: separate mocks and fakes for exception report manager
