+++ packages/exception_report_manager/test/fakes/fake_base_log_message.dart
@@ -1,0 +1,4 @@
+import 'package:log_manager/log_manager.dart';
+import 'package:mocktail/mocktail.dart';
+
+final class FakeBaseLogMessage extends Fake implements BaseLogMessage {}
