+++ packages/exception_report_manager/test/fakes/fake_flutter_error_details.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/foundation.dart';
+import 'package:mocktail/mocktail.dart';
+
+final class FakeFlutterErrorDetails extends Fake
+    implements FlutterErrorDetails {
+  @override
+  String toString({DiagnosticLevel minLevel = DiagnosticLevel.info}) {
+    return 'Fake Flutter Error Details';
+  }
+}
