+++ packages/exception_report_manager/test/mocks/mock_exception_report_manager.dart
@@ -1,0 +1,5 @@
+import 'package:exception_report_manager/exception_report_manager.dart';
+import 'package:mocktail/mocktail.dart';
+
+class MockExceptionReportManager extends Mock
+    implements ExceptionReportManager {}
