+++ packages/exception_report_manager/test/mocks/mock_firebase_crashlytics.dart
@@ -1,0 +1,5 @@
+import 'package:firebase_crashlytics/firebase_crashlytics.dart';
+import 'package:mocktail/mocktail.dart';
+
+final class MockFirebaseCrashlytics extends Mock
+    implements FirebaseCrashlytics {}
