+++ packages/exception_report_manager/test/mocks/mock_log_manager.dart
@@ -1,0 +1,4 @@
+import 'package:log_manager/log_manager.dart';
+import 'package:mocktail/mocktail.dart';
+
+final class MockLogManager extends Mock implements LogManager {}
